feat(nerfstudio_renderer/render_queue.py): Added naive code for serving render requests.
One can register a "render request" to a NerfStudioRenderQueue at any time, and specify a callback function.
To deliver results in time, a render request serially makes multiple rendering calls, starting from the lowest quality.
At the end of each rendering call, an image is obtained, and is passed to the callback function.
The current method will cause many problems, and optimization will be done to counteract them in subsequent commits.